Renamed panther ingest directory from 'orthology' to 'panther' Added pipeline post-processing step to untar & re-gzip the panther download file
