+++ monarch_ingest/panther/ref_genome_orthologs.py
@@ -5,7 +5,7 @@ import uuid
 
 from koza.cli_runner import koza_app
 from biolink_model_pydantic.model import Gene, Predicate, GeneToGeneHomologyAssociation
-from monarch_ingest.orthology.orthology_utils import parse_gene
+from monarch_ingest.panther.orthology_utils import parse_gene
 
 
 row = koza_app.get_row()
@@ -14,13 +14,13 @@ uniprot_2_gene = koza_app.get_map('uniprot_2_gene')
 
 species_and_gene_id = parse_gene(row['Gene'], uniprot_2_gene)
 if species_and_gene_id:
-    
+
     # unpack the species and gene id
     gene_ncbitaxon, gene_id = species_and_gene_id
 
     species_and_ortholog_id = parse_gene(row['Ortholog'], uniprot_2_gene)
     if species_and_ortholog_id:
-        
+
         # unpack the orthogous gene id and its species
         ortholog_ncbitaxon, ortholog_id = species_and_ortholog_id
 
@@ -32,7 +32,7 @@ if species_and_gene_id:
         # build the Gene and Orthologous Gene nodes
         gene = Gene(id=gene_id, in_taxon=gene_ncbitaxon, source="infores:entrez")
         ortholog = Gene(id=ortholog_id, in_taxon=ortholog_ncbitaxon, source="infores:entrez")
-        
+
         # Instantiate the instance of Gene-to-Gene Homology Association
         association = GeneToGeneHomologyAssociation(
             id=f"uuid:{str(uuid.uuid1())}",
@@ -42,6 +42,6 @@ if species_and_gene_id:
             relation=relation,
             source="infores:panther"
         )
-    
+
         # Write the captured Association out
         koza_app.write(gene, ortholog, association)
